+++ java/src/main/java/io/cucumber/tagexpressions/TagExpressionParser.java
@@ -6,6 +6,8 @@ import java.util.Deque;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public final class TagExpressionParser {
     private static final Map<String, Assoc> ASSOC = new HashMap<String, Assoc>() {{
@@ -24,16 +26,16 @@ public final class TagExpressionParser {
     private final String infix;
 
     public static Expression parse(String infix) {
-	return new TagExpressionParser(infix).parse();
+        return new TagExpressionParser(infix).parse();
     }
 
     private TagExpressionParser(String infix) {
-	this.infix = infix;
+        this.infix = infix;
     }
-    
-    private Expression parse() {	
+
+    private Expression parse() {
         List<String> tokens = tokenize(infix);
-        if(tokens.isEmpty()) return new True();
+        if (tokens.isEmpty()) return new True();
 
         Deque<String> operators = new ArrayDeque<>();
         Deque<Expression> expressions = new ArrayDeque<>();
@@ -49,7 +51,7 @@ public final class TagExpressionParser {
                         (ASSOC.get(token) == Assoc.LEFT && PREC.get(token) <= PREC.get(operators.peek()))
                                 ||
                                 (ASSOC.get(token) == Assoc.RIGHT && PREC.get(token) < PREC.get(operators.peek())))
-                        ) {
+                ) {
                     pushExpr(pop(operators), expressions);
                 }
                 operators.push(token);
@@ -64,7 +66,7 @@ public final class TagExpressionParser {
                     pushExpr(pop(operators), expressions);
                 }
                 if (operators.size() == 0) {
-                    throw new TagExpressionException("Tag expression '%s' could not be parsed because of syntax error: unmatched )", this.infix);
+                    throw new TagExpressionException("Tag expression \"%s\" could not be parsed because of syntax error: Unmatched ).", this.infix);
                 }
                 if ("(".equals(operators.peek())) {
                     pop(operators);
@@ -79,7 +81,7 @@ public final class TagExpressionParser {
 
         while (operators.size() > 0) {
             if ("(".equals(operators.peek())) {
-                throw new TagExpressionException("Tag expression '%s' could not be parsed because of syntax error: unmatched (", infix);
+                throw new TagExpressionException("Tag expression \"%s\" could not be parsed because of syntax error: Unmatched (.", infix);
             }
             pushExpr(pop(operators), expressions);
         }
@@ -89,43 +91,32 @@ public final class TagExpressionParser {
 
     private static List<String> tokenize(String expr) {
         List<String> tokens = new ArrayList<>();
-
         boolean isEscaped = false;
-        StringBuilder token = null;
+        StringBuilder token = new StringBuilder();
         for (int i = 0; i < expr.length(); i++) {
             char c = expr.charAt(i);
-            if (ESCAPING_CHAR == c && !isEscaped) {
-                isEscaped = true;
-            } else {
-                if (Character.isWhitespace(c)) { // skip
-                    if (null != token) { // end of token
-                        tokens.add(token.toString());
-                        token = null;
-                    }
+            if (isEscaped) {
+                if (c == '(' || c == ')' || c == '\\' || Character.isWhitespace(c)) {
+                    token.append(c);
+                    isEscaped = false;
                 } else {
-                    switch (c) {
-                        case '(':
-                        case ')':
-                            if (!isEscaped) {
-                                if (null != token) { // end of token
-                                    tokens.add(token.toString());
-                                    token = null;
-                                }
-                                tokens.add(String.valueOf(c));
-                                break;
-                            }
-                        default:
-                            if (null == token) { // start of token
-                                token = new StringBuilder();
-                            }
-                            token.append(c);
-                            break;
-                    }
+                    throw new TagExpressionException("Tag expression \"%s\" could not be parsed because of syntax error: Illegal escape before \"%s\".", expr, c);
                 }
-                isEscaped = false;
+            } else if (c == ESCAPING_CHAR) {
+                isEscaped = true;
+            } else if (c == '(' || c == ')' || Character.isWhitespace(c)) {
+                if (token.length() > 0) {
+                    tokens.add(token.toString());
+                    token = new StringBuilder();
+                }
+                if (!Character.isWhitespace(c)) {
+                    tokens.add(String.valueOf(c));
+                }
+            } else {
+                token.append(c);
             }
         }
-        if (null != token) { // end of token
+        if (token.length() > 0) {
             tokens.add(token.toString());
         }
         return tokens;
@@ -133,12 +124,13 @@ public final class TagExpressionParser {
 
     private void check(TokenType expectedTokenType, TokenType tokenType) {
         if (expectedTokenType != tokenType) {
-            throw new TagExpressionException("Tag expression '%s' could not be parsed because of syntax error: expected %s", infix, expectedTokenType.toString().toLowerCase());
+            throw new TagExpressionException("Tag expression \"%s\" could not be parsed because of syntax error: Expected %s.", infix, expectedTokenType.toString().toLowerCase());
         }
     }
 
     private <T> T pop(Deque<T> stack) {
-        if (stack.isEmpty()) throw new TagExpressionException("Tag expression '%s' could not be parsed because of an empty stack", infix);
+        if (stack.isEmpty())
+            throw new TagExpressionException("Tag expression \"%s\" could not be parsed because of an empty stack", infix);
         return stack.pop();
     }
 
@@ -183,7 +175,7 @@ public final class TagExpressionParser {
         RIGHT
     }
 
-    private class Literal implements Expression {
+    private static class Literal implements Expression {
         private final String value;
 
         Literal(String value) {
@@ -197,11 +189,15 @@ public final class TagExpressionParser {
 
         @Override
         public String toString() {
-            return value.replaceAll("\\\\", "\\\\\\\\").replaceAll("\\(", "\\\\(").replaceAll("\\)", "\\\\)");
+            return value
+                    .replaceAll(Pattern.quote("\\"), Matcher.quoteReplacement("\\\\"))
+                    .replaceAll(Pattern.quote("("), Matcher.quoteReplacement("\\("))
+                    .replaceAll(Pattern.quote(")"), Matcher.quoteReplacement("\\)"))
+                    .replaceAll("\\s", "\\\\ ");
         }
     }
 
-    private class Or implements Expression {
+    private static class Or implements Expression {
         private final Expression left;
         private final Expression right;
 
@@ -221,7 +217,7 @@ public final class TagExpressionParser {
         }
     }
 
-    private class And implements Expression {
+    private static class And implements Expression {
         private final Expression left;
         private final Expression right;
 
@@ -241,7 +237,7 @@ public final class TagExpressionParser {
         }
     }
 
-    private class Not implements Expression {
+    private static class Not implements Expression {
         private final Expression expr;
 
         Not(Expression expr) {
@@ -259,10 +255,15 @@ public final class TagExpressionParser {
         }
     }
 
-    private class True implements Expression {
+    private static class True implements Expression {
         @Override
         public boolean evaluate(List<String> variables) {
             return true;
+        }
+
+        @Override
+        public String toString() {
+            return "true";
         }
     }
 }